chore(www): stop keeping a second set of baselines for CI
The screenshot suite needed a baseline per platform, so every visual
change meant generating the Linux set inside an emulated amd64 container
matching CI's image. That put "can I add a component" behind a local
Docker daemon — which wedged mid-task once already — and it was
undocumented, so an outside contributor following CONTRIBUTING to the
letter would have gotten a red check and no idea why.

It also bought less than it looked like. A baseline catches a change
against the last render somebody accepted; it cannot catch a render that
was wrong from the start. Both kernel bugs fixed in the previous commit
sat inside green baselines for weeks, faithfully preserved as correct.
And the tolerance is wide: 1% of a 663x320 preview is ~2,100 pixels, so
a tooltip tail was added, restyled and removed without ever failing a
test. Tightening it is not available either — measured run-to-run jitter
on one machine, with nothing changed, is 400-1,000 pixels, which is what
that bound is really sized for.

So the suite splits by what each half is actually good at.

checks.spec.ts is the CI gate: console errors on the pages exercising
the awkward patterns, the code-tab geometry invariant, and the
filter-cache one that was previously asserted halfway through a
screenshot test. All direct assertions, no baselines, nothing that
drifts when a browser updates. The workflow is renamed to Checks and no
longer pins its image for pixel fidelity — it just wants a browser.

visual.spec.ts keeps the screenshots and is a local tool. Baselines are
whatever platform you run on; the committed set is macOS, and the Linux
set is deleted rather than left to rot. Run it either side of a change
to the optics and look at what moved.

Both are documented in CONTRIBUTING, which said nothing about visual
tests at all before this.
