+++ .github/workflows/checks.yml
@@ -1,4 +1,4 @@
-name: Visual
+name: Checks
 
 on:
   pull_request:
@@ -6,17 +6,16 @@ on:
     branches: [main]
 
 concurrency:
-  group: visual-${{ github.ref }}
+  group: checks-${{ github.ref }}
   cancel-in-progress: true
 
 jobs:
-  screenshots:
-    name: Glass screenshots
+  e2e:
+    name: Page checks
     runs-on: ubuntu-latest
-    # Pinned to the same image the baselines are generated in. Glass is a
-    # per-pixel canvas render fed through an SVG filter, so it is sensitive to
-    # the graphics stack in a way ordinary DOM screenshots are not — running on
-    # a bare runner would drift from local baselines constantly.
+    # Playwright's own image, purely because it ships the browser and its system
+    # libraries ready to run. Nothing here compares pixels, so the tag no longer
+    # has to match anything — bump it with the dependency.
     container:
       image: mcr.microsoft.com/playwright:v1.62.1-noble
       options: --user 1001
@@ -32,9 +31,12 @@ jobs:
       # its webServer, but that run is cached and this fails fast and legibly
       # if the workspace itself is broken.
       - run: pnpm build
-      - run: pnpm --filter www test:visual
+      # Console errors, layout invariants and filter-cache behaviour. The
+      # screenshot suite (e2e/visual.spec.ts) is a local tool and is not run
+      # here — see CONTRIBUTING.
+      - run: pnpm --filter www test:checks
       - uses: actions/upload-artifact@v5
-        if: ${{ !cancelled() }}
+        if: failure()
         with:
           name: playwright-report
           path: apps/www/playwright-report
